added ->
=== Avoid Interrupts

Due to performance reasons we have remove checking for interrupts in the chronicle queue code, As such we recommend that you avoid using chronicle queue with code that generates interrupts. If you can not avoid generating interrupts then we suggest that you create a separate instance of chronicle-queue per thread.